Adapt events tabs to Wazuh 5.0 schemas (#7934)
* adapt events tabs to Wazuh 5.0 schemas

Replace occurrences of 'rule.groups' with 'wazuh.decoders' across various files to align with the updated schema in Wazuh 5.0. This change ensures compatibility with the new data structure and improves the accuracy of data filtering and visualization in the dashboard components.

* refactor usage of decoders to integration decoders

This commit updates all instances of the variable «wazuh.decoders» to «wazuh.integration.decoders» across various files in the codebase. The change aims to standardize the naming convention for decoder fields, ensuring consistency and clarity in the code.

The updates include changes in TypeScript files, JavaScript files, and test files, affecting both the data source definitions and the alert generation logic. This refactor is crucial for maintaining a clear structure in the code and avoiding confusion regarding the usage of decoder fields in the Wazuh integration context.

* rename «rootcheck» to «wazuh-rootcheck» across codebase

Standardize naming by updating all instances of «rootcheck»
to «wazuh-rootcheck» in decoders, rule mappings, sample data,
dashboard filters, and test logic. Improves clarity and
consistency between data sources and alert generation.

* update constant «CONFIGURATION_ASSESSMENT_VALUE» to use the correct value

This change updates the constant «CONFIGURATION_ASSESSMENT_VALUE» from 'sca' to 'wazuh-sca' to align with the new Wazuh 5.0 schemas. This ensures that the data source correctly references the updated decoder naming convention.

* update constant «FIM_GROUP_VALUE» to use the correct value

This change updates the constant «FIM_GROUP_VALUE» from 'syscheck' to 'wazuh-fim' to align with the new Wazuh 5.0 schemas. This ensures that the data source correctly references the appropriate decoder group for FIM events.

* update constant «AUDITING» to use the correct value

Replaced the value of the constant «AUDITING» from 'audit' to 'azure-audit' to align with the updated schema in Wazuh 5.0. This change ensures consistency across the codebase and reflects the correct auditing module identifier.

* update constant «AZURE_GROUP_VALUE» to use the correct value

This change updates the value of the constant «AZURE_GROUP_VALUE» from 'ms-graph' to 'azure-graphactivity' to align with the Wazuh 5.0 schemas. This ensures that the Azure data source correctly references the intended activity type for better integration and data accuracy.

* update constant «GOOGLE_CLOUD_GROUP_VALUE» to use the correct value

The value for «GOOGLE_CLOUD_GROUP_VALUE» has been updated from 'gcp' to 'gcp-audit' to align with the new Wazuh 5.0 schemas. This change ensures that the Google Cloud integration correctly references the appropriate decoder for audit logs.

* update constant «AWS_GROUP_VALUE» to use the correct value

This change updates the constant «AWS_GROUP_VALUE» from 'amazon' to 'aws-audit' to align with the Wazuh 5.0 schemas. This ensures that the correct value is used in the AWS data source for filtering purposes.

* fix prettier issues

* remove: check-index service file

* remove: unused methods of saved object service

* fix: rootcheck sample data marker

* remove: unused validateIndexPattern from wazuh-elastic controller

* fix: change to wazuh-rootcheck the sample data

* remove: unused files related to dashboard definitions and deprecated dashboards of FIM

* remove: unused files related to dashboard definitions and deprecated dashboards of Malware Detection

* remove: deprecated filter-handler to create filters in favor to PatternDataSourceFilterManager.createFilter

* fix(test): fix tests

* remove: unused dashboard definition creator of FIM Regsitry Keys

* fix: revert change of rule.groups in FIM dashboard

* fix: revert change of rule.groups in GDPR dashboard

* chore(changelog): add entry

* Replace timestamp field for @timestamps field en Events count evolution

* fix: indexer URL replacement in Wazuh server config, add content_manager setting to Wazuh indexer config, generate local in Wazuh server image from package

---------

Co-authored-by: Ian Yenien Serrano <63758389+yenienserrano@users.noreply.github.com>
Co-authored-by: Antonio David Gutiérrez <antonio.gutierrez@wazuh.com>
Co-authored-by: Federico Rodriguez <federico.rodriguez@wazuh.com>