Implement retry logic for yarn install in deployment scripts (#7973)
* implement retry logic for yarn install in deployment scripts

Add a retry mechanism for the «yarn install» command in the deployment scripts of both the base and dev environment workflows. This change aims to enhance the robustness of the installation process by allowing multiple attempts in case of transient failures, thus reducing the likelihood of internal server errors during package fetching.

* add retry logic for yarn install in workflows

Enhance the deployment workflows by adding retry logic for the «yarn install» command. This change introduces an echo statement to indicate the current attempt and captures the exit code if the command fails. This improvement aims to reduce the likelihood of failures during the installation process, providing a more robust deployment experience.

* Testing workflow retry

* Revert "Testing workflow retry"

This reverts commit 0f2c2117ab0c76d326c146be94552894f25dd091.

---------

Co-authored-by: Federico Rodriguez <federico.rodriguez@wazuh.com>