Replace the not run kpi with not applicable kpi (#7979)
* Update SCA dashboards and visualizations to enhance data representation

This commit modifies the SCA inventory and overview dashboards by updating visualizations to include improved color coding for check results and ensuring consistency in the dashboard structure. The changes enhance the clarity and usability of the dashboards, providing better insights into the SCA data. Additionally, the updates align with the latest data handling practices and improve the overall user experience.

* Refactor SCA check results to remove 'Not Run' status and update related components

This commit removes the 'Not Run' check result from the SCA constants and visualizations, replacing it with 'Not Applicable' where necessary. The changes affect various components, including dashboards and visualizations, enhancing clarity and consistency in the representation of SCA results. Additionally, updates to documentation reflect this change, ensuring users have accurate information regarding check results.

* Update CHANGELOG to reflect removal of 'Not Run' status and replacement with 'Not Applicable' in SCA module

This commit updates the CHANGELOG to document the removal of the 'Not Run' check result status from the SCA module, including its replacement with 'Not Applicable' in the SCA Configuration Assessment Dashboard. The changes also address the vertical alignment of the Score KPI, ensuring improved clarity and consistency in the dashboard's representation of SCA results.

* Update CHANGELOG.md

Co-authored-by: Guido Modarelli <38738725+guidomodarelli@users.noreply.github.com>

* Enhance SCA dashboard KPIs by adjusting vertical alignment and adding new visualizations for 'Checks not applicable' and 'Checks applicable'. This update improves the clarity of KPI representation and introduces additional metrics for better data insights.

* Adjust visualizations height

* Fix changelog

---------

Co-authored-by: Guido Modarelli <38738725+guidomodarelli@users.noreply.github.com>
Co-authored-by: Federico Rodriguez <federico.rodriguez@wazuh.com>